Updated controllers

+++ src/main/java/pl/edu/pg/apkademikbackend/commonSpace/exception/CommonSpaceNotFoundException.java
@@ -1,7 +1,7 @@
 package pl.edu.pg.apkademikbackend.commonSpace.exception;
 
 public class CommonSpaceNotFoundException extends RuntimeException {
-    public CommonSpaceNotFoundException(Integer number) {
+    public CommonSpaceNotFoundException(long number) {
         super("Common Space not found " + number);
     }
 }
